reorganising helper files/functions, helper routes file for clients, skeleton tests for client
